requirements.txt
transformers==4.26.0
torch==1.12.0
pillow==9.2.0
pytesseract==0.3.8
streamlit==1.9.0
gradio==2.8.0
